Working version with only nominal value fitted

+++ HistMaker/Hadd.sh
@@ -1,20 +1,28 @@
 #!/bin/sh
 
-p=/eos/user/s/siluo/WPrimeAnalysis/Validation/Hists/
+p=/eos/user/s/siluo/WPrimeAnalysis/ValidationFitted/Hists/
 pre=Validation
-for i in LeptonPt LeptonEta LeptonPhi Jet0Pt Jet0Eta Jet0Phi Jet1Pt Jet1Eta Jet1Phi Jet2Pt Jet2Eta Jet2Phi Jet3Pt Jet3Eta Jet3Phi Jet4Pt Jet4Eta Jet4Phi METPt METPhi dPhiMetLep mT HT ST WPrimeMassSimpleFL WPrimeMassSimpleLL #WPrimeMass WPrimeMassFL WPrimeMassLL Likelihood
+for i in LeptonPt LeptonEta LeptonPhi Jet0Pt Jet0Eta Jet0Phi Jet1Pt Jet1Eta Jet1Phi Jet2Pt Jet2Eta Jet2Phi Jet3Pt Jet3Eta Jet3Phi Jet4Pt Jet4Eta Jet4Phi METPt METPhi dPhiMetLep mT HT ST WPrimeMassSimpleFL WPrimeMassSimpleLL WPrimeMass WPrimeMassFL WPrimeMassLL Likelihood
 # for i in Jet0Eta Jet0Phi Jet1Pt Jet1Eta Jet1Phi Jet2Pt Jet2Eta Jet2Phi Jet3Pt Jet3Eta Jet3Phi Jet4Pt Jet4Eta Jet4Phi METPt METPhi dPhiMetLep mT HT ST WPrimeMassSimpleFL WPrimeMassSimpleLL
 do
-for j in 2016 2017 2018
+for j in 2018
 do
-rm ${p}${j}_${pre}_RW_${i}_ttbar.root
-hadd -f ${p}${j}_${pre}_RW_${i}_ttbar.root ${p}${j}_ttbar/${j}_${pre}_RW_${i}_ttbar_*
-# rm ${p}${j}_${pre}_NRW_${i}_ttbar.root
-# hadd -f ${p}${j}_${pre}_NRW_${i}_ttbar.root ${p}${j}_ttbar/${j}_${pre}_NRW_${i}_ttbar_*
-# rm ${p}${j}_${pre}_${i}_SingleElectron.root
-# hadd -f ${p}${j}_${pre}_${i}_SingleElectron.root ${p}${j}_SingleElectron/${j}_${pre}_${i}_SingleElectron_* 
-# rm ${p}${j}_${pre}_${i}_SingleMuon.root
-# hadd -f ${p}${j}_${pre}_${i}_SingleMuon.root ${p}${j}_SingleMuon/${j}_${pre}_${i}_SingleMuon_* 
+if [ $1 -eq 0 ]; then
+  rm ${p}${j}_${pre}_NRW_${i}_ttbar.root
+  hadd -f ${p}${j}_${pre}_NRW_${i}_ttbar.root ${p}${j}_ttbar/${j}_${pre}_NRW_${i}_ttbar_*
+fi
+if [ $1 -eq 1 ]; then
+  rm ${p}${j}_${pre}_${i}_SingleElectron.root
+  hadd -f ${p}${j}_${pre}_${i}_SingleElectron.root ${p}${j}_SingleElectron/${j}_${pre}_${i}_SingleElectron_* 
+fi
+if [ $1 -eq 2 ]; then
+  rm ${p}${j}_${pre}_${i}_SingleMuon.root
+  hadd -f ${p}${j}_${pre}_${i}_SingleMuon.root ${p}${j}_SingleMuon/${j}_${pre}_${i}_SingleMuon_* 
+fi
+if [ $1 -eq 3 ]; then
+  rm ${p}${j}_${pre}_RW_${i}_ttbar.root
+  hadd -f ${p}${j}_${pre}_RW_${i}_ttbar.root ${p}${j}_ttbar/${j}_${pre}_RW_${i}_ttbar_*
+fi
 
 # rm outputs/${pre}${i}.root
 # hadd -f outputs/${pre}${i}.root ${p}${pre}${i}* ${p}${pre}NRW_${i}*
